mystuff

+++ matlab/trk/trkMovie.m
@@ -1,6 +1,10 @@
-function trkMovie(mv, folder, resultsFolder, filename)
+function trkMovie(mv, folder, resultsFolder, filename, rmFileFlag)
+% folder  where you construct the movie
+% resultsfolder  where the movie ends up
 
-
+if nargin < 5
+    rmFileFlag = 1;
+end
 
 
 disp('...writing temporary image files');
@@ -10,13 +14,7 @@ end
 disp('...encoding movie');
 oldpath = pwd;
 cd(folder);
-%BITRATE = '5000';
-%BITRATE = '25000000';
-%FPS = '10';
-%cmd = ['mencoder "mf://*.png" -mf fps=' FPS ' -o ' resultsFolder filename ' -ovc xvid -xvidencopts bitrate=' BITRATE ' -really-quiet'];
-% cmd = ['ffmpeg -r 10 -b 600k -i %03d.png ' resultsFolder filename];
-%disp(cmd);
-%system(cmd);
+
 
 BITRATE = 15000000;
 cmd1 = ['mencoder -ovc lavc -lavcopts vcodec=msmpeg4v2:vpass=1:"vbitrate=' num2str(BITRATE) ':mbd=2:keyint=132:vqblur=1.0:cmp=2:subcmp=2:dia=2:mv0:last_pred=3" -mf type=png:fps=10 -nosound -o /dev/null mf://*.png -really-quiet'];
@@ -25,18 +23,28 @@ cmd2 = ['mencoder -ovc lavc -lavcopts vcodec=msmpeg4v2:vpass=2:"vbitrate=' num2s
 system(cmd1);
 system(cmd2);
 
+delete([folder 'divx2pass.log']);
 
 cd(oldpath);
 
-cmd = ['rm ' folder '*.png'];
+
+if rmFileFlag
+    cmd = ['rm ' folder '*.png'];  
+    system(cmd);
+end
+
+
+
+
+
+
+%BITRATE = '5000';
+%BITRATE = '25000000';
+%FPS = '10';
+%cmd = ['mencoder "mf://*.png" -mf fps=' FPS ' -o ' resultsFolder filename ' -ovc xvid -xvidencopts bitrate=' BITRATE ' -really-quiet'];
+% cmd = ['ffmpeg -r 10 -b 600k -i %03d.png ' resultsFolder filename];
 %disp(cmd);
-system(cmd);
-
-
-
-
-
-
+%system(cmd);
 
 
 % vidObj = VideoWriter(filename);
